final push

+++ api/requests/delete_privatepdf.js
@@ -3,12 +3,21 @@ const PublicModel = require('../models/publicPdf');
 
 // Route to delete a PDF file by filename
 const privatedelete= async (req, res) => {
-  try {
-    const { file_name ,isAdmin} = req.body;
+try {
+    const { file_name ,isAdmin,userid} = req.body;
 
-    if(isAdmin===false||isAdmin==="false"){
+    if(isAdmin==="user"){
+      const existingPrivatePDF = await PrivateModel.findOne({
+        "pdf.filename": file_name,
+        "userId": userid
+      });
       
-      const deletedPdf = await PrivateModel.deleteOne({ filename:file_name });
+      if (!existingPrivatePDF && !userid){
+        return res.status(404).json({ success: false, message: 'PDF file not found' });
+      }
+
+      
+      const deletedPdf = await PrivateModel.deleteOne({ userId:userid,"pdf.filename":file_name });
       
       if (deletedPdf.deletedCount > 0) {
         return res.status(200).json({ success: true, message: 'PDF file(s) deleted successfully' });
@@ -16,7 +25,7 @@ const privatedelete= async (req, res) => {
         return res.status(404).json({ success: false, message: 'PDF file(s) not found' });
       }
     }else{
-      const deletedPdf = await PublicModel.deleteOne({ filename:file_name });
+      const deletedPdf = await PublicModel.deleteOne({ userId:userid,"pdf.filename":file_name });
       
       if (deletedPdf.deletedCount > 0) {
         return res.status(200).json({ success: true, message: 'PDF file(s) deleted successfully' });
@@ -25,7 +34,8 @@ const privatedelete= async (req, res) => {
       }
 
     }
-    } catch (error) {
+    }
+     catch (error) {
       console.error('Error deleting PDF file:', error);
       return res.status(500).json({ success: false, message: 'Internal server error' });
     }
